fix: ready doesn't get cleared after execution
The expected behavior is that the ready function only gets executed once
and then its listener gets cleared from the library after it is
executed.

The current code instead of making a declaration to null its doing a
comparation to null.

This commit changes the behavior of the library for the ready event.